# agrotech/api-spec.yaml
{"openapi":"3.0.1","info":{"title":"Swagger Demo","description":"Documentation APIs v1.0","version":"1.0"},"servers":[{"url":"https://localhost:8454","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/cadastru":{"get":{"tags":["cadastru-controller"],"operationId":"getCadastres","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Cadastru"}}}}}}},"post":{"tags":["cadastru-controller"],"operationId":"saveCadastru","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Cadastru"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Cadastru"}}}}}}},"/user/{id}":{"get":{"tags":["user-controller"],"operationId":"getCadastru","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Cadastru"}}}}}}}},"/raion/{id}":{"get":{"tags":["raion-controller"],"operationId":"getRaion","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Raion"}}}}}}},"/cadastru/{id}":{"get":{"tags":["cadastru-controller"],"operationId":"getCadastru_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Cadastru"}}}}}}},"/cadastru/{id}/livestock":{"get":{"tags":["cadastru-controller"],"operationId":"getLivestock","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LivestockDTO"}}}}}}}},"/cadastru/{id}/crops":{"get":{"tags":["cadastru-controller"],"operationId":"getCrops","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CropDTO"}}}}}}}},"/cadastru/livestock":{"get":{"tags":["cadastru-controller"],"operationId":"getCadastruByLivestock","parameters":[{"name":"livestock","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CadastruLivestockDTO"}}}}}}}},"/cadastru/crop":{"get":{"tags":["cadastru-controller"],"operationId":"getCadastruByCrop","parameters":[{"name":"crop","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CadastruCropDTO"}}}}}}}}},"components":{"schemas":{"Cadastru":{"type":"object","properties":{"cadastruId":{"type":"integer","format":"int64"},"userId":{"type":"string"},"raionId":{"type":"integer","format":"int64"},"hectare":{"type":"integer","format":"int64"},"longitude":{"type":"number","format":"double"},"latitude":{"type":"number","format":"double"}}},"Raion":{"type":"object","properties":{"raionId":{"type":"integer","format":"int64"},"raionName":{"type":"string"},"horizonAhDepth":{"type":"integer","format":"int64"}}},"LivestockDTO":{"type":"object","properties":{"livestockId":{"type":"integer","format":"int64"},"livestockName":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"contaminated":{"type":"number","format":"double"},"dataProduced":{"type":"string","format":"date-time"}}},"CropDTO":{"type":"object","properties":{"cropId":{"type":"integer","format":"int64"},"cropName":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"pesticide":{"type":"number","format":"double"},"dataProduced":{"type":"string","format":"date-time"}}},"CadastruLivestockDTO":{"type":"object","properties":{"cadastruId":{"type":"integer","format":"int64"},"userId":{"type":"string"},"raionId":{"type":"integer","format":"int64"},"hectare":{"type":"integer","format":"int64"},"longitude":{"type":"number","format":"double"},"latitude":{"type":"number","format":"double"},"livestockName":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"contaminated":{"type":"number","format":"double"},"dataProduced":{"type":"string","format":"date-time"}}},"CadastruCropDTO":{"type":"object","properties":{"cadastruId":{"type":"integer","format":"int64"},"userId":{"type":"string"},"raionId":{"type":"integer","format":"int64"},"hectare":{"type":"integer","format":"int64"},"longitude":{"type":"number","format":"double"},"latitude":{"type":"number","format":"double"},"cropName":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"pesticide":{"type":"number","format":"double"},"dataProduced":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}