# agrotech/src/api/apiSchemas.ts
/**
 * Generated by @openapi-codegen
 *
 * @version 1.0
 */
export type Cadastru = {
  /**
   * @format int64
   */
  cadastruId?: number;
  userId?: string;
  /**
   * @format int64
   */
  raionId?: number;
  /**
   * @format int64
   */
  hectare?: number;
  /**
   * @format double
   */
  longitude?: number;
  /**
   * @format double
   */
  latitude?: number;
};

export type Raion = {
  /**
   * @format int64
   */
  raionId?: number;
  raionName?: string;
  /**
   * @format int64
   */
  horizonAhDepth?: number;
};

export type LivestockDTO = {
  /**
   * @format int64
   */
  livestockId?: number;
  livestockName?: string;
  /**
   * @format int32
   */
  quantity?: number;
  /**
   * @format double
   */
  contaminated?: number;
  /**
   * @format date-time
   */
  dataProduced?: string;
};

export type CropDTO = {
  /**
   * @format int64
   */
  cropId?: number;
  cropName?: string;
  /**
   * @format int32
   */
  quantity?: number;
  /**
   * @format double
   */
  pesticide?: number;
  /**
   * @format date-time
   */
  dataProduced?: string;
};

export type CadastruLivestockDTO = {
  /**
   * @format int64
   */
  cadastruId?: number;
  userId?: string;
  /**
   * @format int64
   */
  raionId?: number;
  /**
   * @format int64
   */
  hectare?: number;
  /**
   * @format double
   */
  longitude?: number;
  /**
   * @format double
   */
  latitude?: number;
  livestockName?: string;
  /**
   * @format int32
   */
  quantity?: number;
  /**
   * @format double
   */
  contaminated?: number;
  /**
   * @format date-time
   */
  dataProduced?: string;
};

export type CadastruCropDTO = {
  /**
   * @format int64
   */
  cadastruId?: number;
  userId?: string;
  /**
   * @format int64
   */
  raionId?: number;
  /**
   * @format int64
   */
  hectare?: number;
  /**
   * @format double
   */
  longitude?: number;
  /**
   * @format double
   */
  latitude?: number;
  cropName?: string;
  /**
   * @format int32
   */
  quantity?: number;
  /**
   * @format double
   */
  pesticide?: number;
  /**
   * @format date-time
   */
  dataProduced?: string;
};