# agrotech/src/services/openapi/models/Cadastru.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Cadastru = {
    cadastruId?: number;
    userId?: string;
    raionId?: number;
    hectare?: number;
    longitude?: number;
    latitude?: number;
};