# java/agrotech/src/main/java/md/da/agrotech/DTO/CropDTO.java
package md.da.agrotech.DTO;

import java.time.LocalDateTime;

public record CropDTO(
    Long cropId,
    String cropName,
    Integer quantity,
    Double pesticide,
    LocalDateTime dataProduced
) {}